src/creative_generator/requirements.txt
# Core dependencies
motor==3.3.2  # Compatible with pymongo 4.5.0
python-dotenv==1.0.0
nats-py==2.6.0
protobuf>=3.20.0,<4.0.0  # Version compatible with the generated protobuf code
openai>=1.0.0,<1.1.0  # Use a version compatible with our dependencies
pymongo==4.5.0  # Explicitly set compatible version
httpx>=0.24.0,<0.25.0  # Explicitly set httpx version to avoid compatibility issues

# Development
pytest==7.4.0
pytest-asyncio==0.21.1
